Move code to core module
