+++ patterns-ts/src/index.ts
@@ -1,29 +1,7 @@
-export interface NodeCategory {
-  id: number;
-  name: string;
-  isBuildingBlock: boolean;
-}
+/**
+ * Top-level exports
+ * 
+ * Allow type names from core to be imported from the package root.
+ */
 
-
-export interface RoleCategory extends NodeCategory {
-}
-
-
-export interface ContextCategory extends NodeCategory {
-  iconName: string;
-}
-
-
-export class NodeDefinition {
-  id = 0;
-  name = '';
-  iconName = '';
-  explanation = '';
-  hyphenatedName = '';
-
-  constructor(
-    public contextCat: ContextCategory,
-    public roleCat: RoleCategory,
-    public isCustom = true,
-  ) {}
-}
+export * from './core';
